# api/openapi.yaml
openapi: 3.1.0
info:
  title: NIRVAPROCURE API — Phase 1
  version: 0.1.0
  description: |
    Phase 1 endpoints covering purchase requests, marketplace link import,
    approvals, and LINE notification wiring. All endpoints are scoped to
    the caller's organization via the JWT `org_id` claim.

servers:
  - url: https://api.nirvaprocure.com/v1

security:
  - bearerAuth: []

tags:
  - name: PR
    description: Purchase requests
  - name: Approvals
    description: Approval workflows and decisions
  - name: Notifications
    description: LINE / email / in-app delivery

paths:
  /pr:
    get:
      tags: [PR]
      summary: List purchase requests
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/PrStatus' }
        - in: query
          name: requester_id
          schema: { type: string, format: uuid }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PurchaseRequest' }
                  next_cursor: { type: string, nullable: true }
    post:
      tags: [PR]
      summary: Create a purchase request (draft)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PrCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PurchaseRequest' }

  /pr/import-link:
    post:
      tags: [PR]
      summary: Parse a marketplace product URL into a PR line-item draft
      description: |
        Accepts a Shopee, Lazada, Alibaba, or Makro product URL. Returns a
        line-item draft with description, unit price, supplier name, and
        parsed metadata. The client decides whether to attach it to an
        existing PR or create a new one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri }
      responses:
        '200':
          description: Parsed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParsedLineItem' }
        '422':
          description: URL was reachable but could not be parsed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /pr/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [PR]
      summary: Get a single PR with items and approval trail
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PurchaseRequestDetail' }
    patch:
      tags: [PR]
      summary: Update PR fields or line items (draft only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PrUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PurchaseRequest' }

  /pr/{id}/submit:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [PR]
      summary: Submit a draft PR for approval
      responses:
        '200':
          description: Submitted; approval instance created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PurchaseRequestDetail' }
        '409':
          description: PR not in draft status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /approvals/inbox:
    get:
      tags: [Approvals]
      summary: List PRs currently awaiting the caller's decision
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ApprovalInboxItem' }

  /approvals/{instance_id}/decision:
    parameters:
      - in: path
        name: instance_id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Approvals]
      summary: Approve or reject at the caller's current step
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approved, rejected] }
                comment:  { type: string }
      responses:
        '200':
          description: Decision recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PurchaseRequestDetail' }

  /notifications/line/test:
    post:
      tags: [Notifications]
      summary: Send a test LINE notification to the caller
      responses:
        '202':
          description: Queued

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    PrStatus:
      type: string
      enum: [draft, submitted, in_approval, approved, rejected, cancelled, completed]

    Money:
      type: object
      properties:
        amount_minor: { type: integer, description: "Amount in minor units (e.g. satang)" }
        currency:     { type: string, example: THB }
      required: [amount_minor, currency]

    PrCreate:
      type: object
      required: [title, items]
      properties:
        title:          { type: string }
        justification:  { type: string }
        department_id:  { type: string, format: uuid }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/PrLineItemInput' }

    PrUpdate:
      type: object
      properties:
        title:         { type: string }
        justification: { type: string }
        department_id: { type: string, format: uuid }
        items:
          type: array
          items: { $ref: '#/components/schemas/PrLineItemInput' }

    PrLineItemInput:
      type: object
      required: [description, quantity, unit_price_minor]
      properties:
        description:      { type: string }
        quantity:         { type: number }
        unit:             { type: string, example: unit }
        unit_price_minor: { type: integer }
        supplier_id:      { type: string, format: uuid }
        source:           { type: string, enum: [shopee, lazada, alibaba, makro, manual] }
        source_url:       { type: string, format: uri }
        source_metadata:  { type: object, additionalProperties: true }

    PurchaseRequest:
      type: object
      properties:
        id:            { type: string, format: uuid }
        pr_number:     { type: string, example: "PR-2026-0001" }
        title:         { type: string }
        status:        { $ref: '#/components/schemas/PrStatus' }
        requester_id:  { type: string, format: uuid }
        department_id: { type: string, format: uuid, nullable: true }
        total:         { $ref: '#/components/schemas/Money' }
        submitted_at:  { type: string, format: date-time, nullable: true }
        created_at:    { type: string, format: date-time }

    PurchaseRequestDetail:
      allOf:
        - $ref: '#/components/schemas/PurchaseRequest'
        - type: object
          properties:
            justification: { type: string }
            items:
              type: array
              items: { $ref: '#/components/schemas/PrLineItem' }
            approval:
              $ref: '#/components/schemas/ApprovalTrail'

    PrLineItem:
      type: object
      properties:
        id:               { type: string, format: uuid }
        line_no:          { type: integer }
        description:      { type: string }
        quantity:         { type: number }
        unit:             { type: string }
        unit_price_minor: { type: integer }
        line_total_minor: { type: integer }
        supplier_id:      { type: string, format: uuid, nullable: true }
        source:           { type: string, nullable: true }
        source_url:       { type: string, nullable: true }

    ParsedLineItem:
      type: object
      properties:
        source:           { type: string, enum: [shopee, lazada, alibaba, makro] }
        source_url:       { type: string, format: uri }
        description:      { type: string }
        unit_price_minor: { type: integer }
        currency:         { type: string, example: THB }
        supplier:
          type: object
          properties:
            name:         { type: string }
            external_ref: { type: string }
        source_metadata:  { type: object, additionalProperties: true }

    ApprovalInboxItem:
      type: object
      properties:
        instance_id:    { type: string, format: uuid }
        pr:             { $ref: '#/components/schemas/PurchaseRequest' }
        step_no:        { type: integer }
        waiting_since:  { type: string, format: date-time }

    ApprovalTrail:
      type: object
      properties:
        instance_id:   { type: string, format: uuid }
        current_step:  { type: integer }
        status:        { type: string, enum: [pending, approved, rejected, cancelled] }
        decisions:
          type: array
          items:
            type: object
            properties:
              step_no:     { type: integer }
              approver_id: { type: string, format: uuid }
              decision:    { type: string, enum: [approved, rejected] }
              comment:     { type: string }
              decided_at:  { type: string, format: date-time }

    Error:
      type: object
      required: [code, message]
      properties:
        code:    { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }